Function rstack_self::trace

pub fn trace(child: &mut Command) -> Result<Vec<Thread>>

Returns stack traces of all of the threads the calling process.

The provided Command should be configured to spawn a process which will call the child function. It must not use standard input or standard output, but standard error will be inherited and can be used. The spawned process must "directly" call child rather than spawning another process to call it. That is, the parent of the process that calls child is the one that will be traced.